requirements.txt
requests~=2.25.1
matplotlib~=3.3.4
pillow~=8.2.0
python-dotenv~=0.18.0
cffi~=1.14.5
future~=0.18.2
ipython~=7.22.0
zlib~=1.2.11
olefile~=0.46
pip~=21.0.1
wheel~=0.36.2
openssl~=1.1.1k
cryptography~=3.4.7
py~=1.10.0
lxml~=4.6.3
tornado~=6.1
keyring~=22.3.0
pytest~=6.2.3
numpy~=1.20.1
cython~=0.29.23
setuptools~=52.0.0
pytz~=2021.1
nose~=1.3.7
numba~=0.53.1
typing_extensions~=3.7.4.3
pkginfo~=1.7.0
click~=7.1.2
certifi~=2020.12.5
chardet~=4.0.0
urllib3~=1.26.4
six~=1.15.0
python-dateutil~=2.8.1
idna~=2.10
cycler~=0.10.0
ipykernel~=5.3.4
path~=15.1.2
sphinx~=4.0.1
docutils~=0.17
jinja2~=2.11.3
pyparsing~=2.4.7
kiwisolver~=1.3.1
